mcount/x86: Implement dynamic update for fentry
The gcc on x86_64 support -mfentry and -mnop-mcount option to generate
instrumentation code replaced with nop.  This can reduce profile
overhead when not enabled as they are nop.  But profiler should patch
the code to replace the nop into the original calls.

This patch implements basic support for the dynamic patching all
functions in a main executable.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>