record: Use epoll to wait available data
With kernel tracing, a writer thread should wait data from multiple
sources.  It'd be much efficient to wait using epoll instead of a loop
with arbitrary timeouts.

To do this, it changes to use a eventfd to signal the writer threads
rather than the pthread conditional variable.  By using a fd, it's easy
to wait on the buffer using epoll.

This change reduces number of LOST messages in kernel significantly.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>